The Great Grass Debacle Is Finished
Ryan and I solved a HUGE problem. What the scale of our project would be. We discussed how large we would want the grass in the game to be, and it was hard to determine what we wanted. We finally decided on a scale to draw Ryan's assets in so that they can be scaled down. We got the grass into the game and it looks great. We also agreed that it might be beneficial to have the grass all be separately drawn objects in the game. This is because each grass object could have a random chance of spawning a pokemon and each one would need to trigger an animation for leaves to crumble and for rustling noises to play. Unsure if having that many separate objects might make the game unstable, but only one way to find out!