Adjusted Grass Image
Adjusted the dimensions of the grass image from 500x500 to 180x180 pixels because that the 180z180 size allows for 10 patches of grass from top to bottom, and about 20 patches of grass horizontally from left to right.

Also figured out how to properly scale both up and down without quality loss in Gimp. All we had to do was turn off interpolation.