Camera Movement (Screen Translation) Completed
Fixed up the code so that the screen translates properly based on the characters movements (camera follows character, but prevents character from moving past the edges of the level)

Next up I plan to utilize tiles and the mapmaking concept I learned in class, where the numbers in a table correspond to the type of tile to draw. In this case, however, I will multi-purpose it to work with simple scenery tiles with animations and other ideas.

I am going to try to create all of the physics at the same time as I'm building the tile system. It's gonna be WILD!!