Started Tilemap Functionality + Updated ClearMeadowTown.xcf
I thought tilemaps would be easy, but with various assets at various different sizes, I was struggling with how to go about doing it.

I decided I'd make it a one size fits all scale for each scene. This first one, Clear Meadow Town, has a bit of an odd tile size, but I think it'll be better going forward. I need to think of the best way to do this, and I'm not sure if my current idea is it.

I also added all of the layers for houses, flowers, bench, lab, trees, etc. into the ClearMeadowTown.xcf file.