Added Individual Files for Different Scenes/Levels
Created a lua file for clearMeadowTown, and I'm setting up all of the relevant data/code for that scene inside of the file. After it's all set, I'll import it into the main.lua file and use it.

I plan to do this for all Scenes going forward, as I think it is going to help with keeping things organized and this will make the most sense.