Updated Folder Structure and New Tilemap Verdict
I made some modifications to the assets structure so that it is more organized.

I also have come to the conclusion that the reason the tilemap implementation has been so difficult is because it shouldn't really be implemented. I don't think it works as well in a top-down game, and there are too many issues with scale for it to be worthwhile.

Additionally, I won't really be needing to draw that many things. At most I'll be drawing a few flowers and some other decorative assets in each level, and it won't be hard to manually code the positions for those assets.