Modified Scaling
In-game, the character looks really good. I'm happy with his dimensions and I think we have a good speed set for him too.

The issue, however, is that the other assets in Clear Meadow Town have been scaled too large in respect to the character. Since I'm happy with the character, I don't want to mess with his dimensions! I'm scaling everything else down to the character's size.

I updated the scale sizes of:
- The houses
- The lab
- The horizontal fence
- The vertical fence
- The bench
- The flowers

They are now all in appropriate sizes proportionate to the character.