More Scaling and Physics Debugging
I made some more scaling modifications and got all of the physics objects to have the proper dimensions.

Once everything was put onto the screen, it looked really good, but the collision was all sorts of messed up. I'm not really sure what the issue is, and I played around for about an hour trying to figure it out.

I'm taking a break now, but I'll be working on that next.