Lots of Physics Work
I've spent a long while today on physics. I figured out that the shape acts as a sort of hitbox. Even if you set the width and height of the shape to 0 and 0, there will still be a 1 px hitbox.

Anyways, I tweaked physics.lua so that the addObject function takes in hitboxX, hitboxY, hitboxWidth, and hitboxHeight as arguments too. The function now automatically positions and sets the hitbox up.

I also included a new system that helps determine where to draw objects based on the position of the player. Basically in the clearMeadowTown.lua file, I created two lists in the clearMeadowTown.load() function; one to store any objects that will be drawn underneath the player (drawn before the player) and another to store any objects that will be drawn above the player (drawn after the player).

Each object now has a setDrawPosition() function that needs to be manually defined by the user. That function should add an object to either the drawAbovePlayer or drawBelowPlayer list based on the code.

In clearMeadowTown.draw(), a for loop iterates over all the objects in the drawUnderneathPlayer list and calls each object's draw function. Then the player's draw function is called, and after that a separate loop iterates over all the objects in drawAbovePlayer and calls each object's draw function.

This allows me to make the player appear to be moving behind or in front of objects on the screen.