Physics In a Good Place and Tilemap Functional
Physics are doing great, collision detection is working very well and everything seems to be cooperating.

I was able to get all of the major Clear Meadow Town assets on the screen. I created a tilemap for the scene where each tile was 60 px by 60 px, and that fit into the size of 3840x2160px for Clear Meadow Town.

I chose that size because I figured that the assets would all be able to fit reasonably into that size, and that I didn't necessarily need to use all the tiles. I wanted the tile size smaller because I figured I'd rather have more variation and decision over where things move. With a larger size, I'm limited to how close I can put assets together.

Physics are in a good place right now, and next I'm going to work with quads. I'm going to see if I can make my animation system work on spritesheets instead of individual images, and that will rely on quads. Additionally, I want to use quads to draw two separate halves of an image (grass example) so that I can let the player interact with objects in a multi-dimensional manner.