Diagonal Movement and Animation Speed Bug
Diagonal movement is implemented and functioning very nicely. I might rework the code to be a bit less redundant at some point, but not a concern for the moment. I'm not using diagonal animation frames, as those were too difficult to draw for a 2D game.

There was a bug I noticed with the speed of animations. No matter what value is passed for the speed argument in the animator.create() function, the speed wouldn't be modified. I fixed it so that the speed works as originally intended to (entering 0.5 will cut the speed in half, 1 will keep the speed the same, 2 will double the speed, and so on)