Animations Split Into Top and Bottom Segments
Using quads to save a top half of an animation frame and a bottom half of an animation frame so that they can be drawn on different layers.

In animator.lua I made it so that the animation stores two separate lists, one holding top half of frames and the other holding the bottom half of frames.

I added functions to draw the top half or bottom half individually for an animation, and I update the default draw function so that it draws both halves at the same time.

Both animator.lua and player.lua are clean and organized at the moment.