Clear Meadow Assets and Spritesheet Bleeding Bug
Clear Meadow Assets are mostly implemented. I think I'm going to rework it so that each object has its own file. Trying to do that currently starting with the shortFlower, but I'm having an issue getting it to draw.

There was a bleeding issue with the spritesheets, but I was able to solve it. I added a 1px transparent border around each sprite in the sheet. It's quite a hassle to do that, but until I can find or think of a better way to make the sheets we're stuck with it.