More Spritesheet and Asset Implementation Work
I finally fixed that weird bug from yesterday. The shortFlower.lua file I made had overwritten the typical object.update() method, which was perfectly fine and what I wanted. But I had forgotten to make sure that duplicates got the overwritten version, which was responsible for drawing the animation, so that's why they wouldn't appear on the screen even though they'd load fine.

Anyways, I figured it out and came up with a better system for the update functions. I added an empty object.customUpdate() function so that every object has it. Then, I made sure that the default update() method in each object calls that customUpdate() method. This allows me to simply add any updates (like player's move() function) by calling the code inside of the customUpdate() method.

I was working with spritesheets, trying to find a consistent and effective way for Ryan to create them for me. It took a while, but I figured out a good way to do it with a website I found. I typed out instructions for Ryan and made modifications to my animation code so that the spritesheets won't bleed. It is a very easy process now.