HUGE Changes to physics.lua
I revamped physics.lua because it had gotten too messy and too difficult to work with. I tidied it up a bit, renamed it to object, and made things a lot more clear I think. Currently it seems almost functional, the only issue is things aren't drawing on the screen. But I'm gonna fix that soon.