Tweaks to Code, Game is In Operable State
I found some errors from my changes in the previous commit and fixed all of them, so the game is running fine now.

I need to get the rest of the Clear Meadow Town assets drawn on the screen, and for some I'll have to make their own files.

I don't want to overdo it with the number of files, so I'll only put universal objects in their own file (that meaning objects that I think would appear in several levels). Objects that might be native to only one level can just be defined in the file.