Clear Meadow Town All Assets Implemented and File Hierarchy Updates
After hours and hours of bugs, technical decisions, stupid mistakes, and a lot of dedication, all of the assets for Clear Meadow Town are implemented and they look really great.

I struggled with this for so long because I was so concerned about how to make my code the most efficient and effective it could be. That killed a lot of time, but I think that time was well spent. Because I spent as long as I did on making sure it'd be convenient, asset implementation should be a much more efficient process going forward.

Also, I updated the file system so that all of the lua scripts are stored in one folder, and inside that scripts folder there are subfolders for objects and scenes. I think this will help with organization, and I've already made the necessary modifications in code.

Scary to think that I haven't even really started the pokemon aspect of this pokemon game. I'm running out of time, just under a month to go. There are some things I feel like I absolutely need to include, and those things need to be worked on first.

I'll need to find a way to easily load in the pokemon for starters, and that might take some time but again, I think it will be worthwhile. After that, I'll need to make an inventory system for the player's bag and pokemon. That means I'll need to make items that the player can use. Also, I'll of course need to make a battling system and implement throwing pokeballs and catching pokemon.

Later on down the line, I want to implement NPC characters. That seems like it might be a bit too much to handle right now, and I don't think I'd be able to get it done before the project is due for school. Also I want to figure out a good way to save progress to a file.