Framework for battler.lua
I started to translate some of my pseudocode into real code for the battler.lua file. I have a really good blueprint for how this battle system should work, and it really didn't even take all that long.

I spent a few hours on it and I think that is time well spent. I think it is much better for me to have planned out the idea first instead of jumping right into it, because I'm figuring out the best way to organize things early on instead of having to rearrange things later and waste time.