Scene Unloading & Attempts To Enter Buildings Smoothly
I modified the scene.lua file so that it contains an unload function. When this function is called, the scene will be unloaded by removing/destroying each component it has (cameras, objects, etc.). Then, if a Scene is loaded again, all of those components are reloaded into the Scene as well.

While implementing the Scene load/unload, I ended up producing a bug with the player. When the Scene changes, the player becomes unable to move. The running animations play, but the position doesn't update and it looks as if the player is running into something. This will need to be fixed