Worked on Pokemon, Textbox, and Battler Functionality
Made some improvements to the Pokemon file. Animations are now working for the Pokemon, and I was able to make textboxes. I'd like for each character to appear one by one really quickly on the textbox if possible, but that's something that I should look into only if I have extra time. The textboxes lift on and fall off of the screen when they're displayed or hidden.

The Battler file is coming along nicely too, I'd say its about a quarter of the way done- maybe a little bit more than that.