Messy Battle Stuff
Still working on the battle system. It was not nearly as simple as I thought it'd be, I'm struggling quite a bit with it.

I no longer have time for a lot of the ideas I'd wanted to implement unfortunately. I have to make the next area with tall grass and I think that's all  I'll be able to do. I wanted to add items like potions, a whole other town with a pokemon center and shop, and I wanted to implement catching pokemon functionality and so much more but things just didn't work out that way.

I at the very least need to have the battling functional, since that will be the main focus point of my game now.